[SIX-276] refactor & fix : 도메인 이름 변경 및 이벤트 핸들러 수정
