+++ onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/notification/NotificationController.java
@@ -1,9 +1,9 @@
-package com.sixheroes.onedayheroapi.alarm;
+package com.sixheroes.onedayheroapi.notification;
 
 import com.sixheroes.onedayheroapi.global.argumentsresolver.authuser.AuthUser;
 import com.sixheroes.onedayheroapi.global.response.ApiResponse;
-import com.sixheroes.onedayheroapplication.alarm.AlarmService;
-import com.sixheroes.onedayheroapplication.alarm.response.AlarmResponse;
+import com.sixheroes.onedayheroapplication.notification.NotificationService;
+import com.sixheroes.onedayheroapplication.notification.response.AlarmResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -15,16 +15,16 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/alarms")
 @RestController
-public class AlarmController {
+public class NotificationController {
 
-    private final AlarmService alarmService;
+    private final NotificationService notificationService;
 
     @GetMapping
     public ResponseEntity<ApiResponse<Slice<AlarmResponse>>> findAlarm(
         @AuthUser Long userId,
         @PageableDefault Pageable pageable
     ) {
-        var alarm = alarmService.findAlarm(userId, pageable);
+        var alarm = notificationService.findAlarm(userId, pageable);
 
         return ResponseEntity.ok(ApiResponse.ok(alarm));
     }
@@ -34,7 +34,7 @@ public class AlarmController {
         @AuthUser Long userId,
         @PathVariable String alarmId
     ) {
-        alarmService.deleteAlarm(userId, alarmId);
+        notificationService.deleteAlarm(userId, alarmId);
 
         return new ResponseEntity<>(ApiResponse.noContent(), HttpStatus.NO_CONTENT);
     }
