+++ onedayhero-api/src/main/java/com/sixheroes/onedayheroapi/sse/SseEventListener.java
@@ -1,10 +1,9 @@
 package com.sixheroes.onedayheroapi.sse;
 
-import com.sixheroes.onedayheroapplication.alarm.dto.SsePayload;
+import com.sixheroes.onedayheroapplication.notification.dto.SsePayload;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.event.TransactionPhase;
-import org.springframework.transaction.event.TransactionalEventListener;
 
 @RequiredArgsConstructor
 @Component
@@ -12,7 +11,7 @@ public class SseEventListener {
 
     private final SseEmitters sseEmitters;
 
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    @EventListener
     public void sendSseEmitter(SsePayload ssePaylod) {
         sseEmitters.send(
             ssePaylod.userId(),
