+++ onedayhero-application/src/main/java/com/sixheroes/onedayheroapplication/notification/dto/SsePayload.java
@@ -1,4 +1,4 @@
-package com.sixheroes.onedayheroapplication.alarm.dto;
+package com.sixheroes.onedayheroapplication.notification.dto;
 
 import com.sixheroes.onedayheromongo.alarm.Alarm;
 import lombok.Builder;
